requirements2.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.2
contourpy==1.3.2
-e git+https://github.com/kwat0308/cr-pulse-interpolator.git@25c6f0e9748ee744629fff4c79a4d59a1338ee5a#egg=cr_pulse_interpolator
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.58.4
h5py==3.14.0
ipykernel==6.29.5
ipython==8.37.0
jax==0.4.38
jax_radio_tools==0.1.3
jaxlib==0.4.38
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
nest-asyncio==1.6.0
nifty==9.1.0
numpy==2.2.6
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyzmq==27.0.0
radiotools==0.2.5
scipy==1.15.3
six==1.17.0
-e git+https://gitlab.iap.kit.edu/AirShowerPhysics/smiet-cr-synthesis.git@3c415058cfbf97d8bfb56824e9e6724a8de0fef2#egg=smiet
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
wcwidth==0.2.13
